separated chat functionality into ChatRoom component
